* proc st test

+++ include/DLWAssembly/ASTypes.h
@@ -9,6 +9,6 @@
 typedef  uint8_t ASByte;
 typedef uint16_t ASWord;
 typedef uint32_t ASDword;
-typedef  int32_t ASInteger;
+typedef  int16_t ASInteger;
 typedef  uint8_t ASSize;
 typedef  uint8_t ASIndex;
